define array type wrappers
